Bugfix: Unwrap Name from a list
Name was getting wrapped in a list since start_link passes a list of
arguments and this was not being unwrapped in the signature of init.

Also, I am told it is a good idea to place guards for your types on
all your functions. I find this a nice idea but it also appears to
make my code longer and polluted in terms of syntax. I add this on one
of the functions as an example and will give this some more thought
later on.

Also, metadata can be declared in a single line.