Add tests for follower and fix bugs
Tests for the follower callback when receiving vote requests and
votes. This exposed a bug where the follower was voting if the term
was equal to its own term. Fixed that.

The fix also makes the should_step_down/2 redundant and will be
removed in a subsequent patch.

Currently the way the tests have been setup, testing to ensure if a
vote was sent is not possible. This is okay for now since we are
working on the unit tests at the moment and will work on the
integration tests later.