Remove hardocded list of three nodes
The leader election now works with five nodes as well and should work
with more. It was stupid to have hardcoded the list, but I am glad it
was just this and not something much more messier. Also defined NODES
as a macro that can be manipulated separately for tests and
production.