Remove hardocded timeout
The random distribution timeout was hardcoded. Moved it to a macro
instead and added docs. The test for upper limit of the timeout was
incorrect as well since the rand:uniform(N) returns a value X, in the
range 1 <= X =< N and our test was using the less than check. This
should fix occasional test failures like the one in this build:
https://travis-ci.org/indradhanush/raft/builds/260330352