Add callbacks for "call" requests from client
This is first of a series of commits and contains expected behaviour
of the follower and candidate when they receive a client request. They
reply with `{error, LeaderId}`. The leader replies with just the `{ok,
Name}` for now as a starting point. This will obviously expand more.

Added tests for each callback as well.