Add a client interface to interact with the cluster
Added a function handle_event/3 to handle an event named test_timeout
which is used in the tests for determinism. This is not the best
approach but the only one that I can think of right now that allows us
to move forward. Also using a higher timeout seed in tests while using
zero in production.