Initialize next_index if new leader gets elected
The leader(timeout,...) function is called when a candidate promotes
itself to leader. Thus we need to initialize next_index for each of
the nodes here as well.

When the leader times out, it will send heartbeats, but chances are
that it has also updated it's log. As a result updating the next_index
to be the