Leader should expect cast from clients
The client should implement its own lock step, by first making a cast
request and then listening for an incoming cast request. We do this
since we want our nodes to not get blocked by a call request from a
client while it sends out RPCs to all neighbouring nodes and waits to
hear from them.

The following will be done in a subsequent commit:
1. Make follower and candidate expect cast for client_messages as
   well.
2. Fix tests.
3. Implement a lock step client API.