TIMEOUT_SEED should not be set to 0
Setting it to 0 means that the election timeout can be closer to the
heartbeat timeout. This makes the followers timeout too frequently and
start a new election. The cluster won't be able to function that way.

Log statement in follower's handler for vote request was
misleading. Fixed it as well in this commit.