raft_client:write will wait for a response
Instead of using a gen_statem:call, the function does a
gen_statem:cast and then immediately goes into a `receive` block
expecting a response. This adds a request-response lock-step
abstraction.

As a result of this, the leader sends a message for the lock-step to
complete on the client side once it has sent out the append entries
RPC request to all nodes.

This commit also adds placeholder function clauses for follower and
candidate to handle append entries with non empty log. This was caught
while writing tests.

The tests in raft_client have been modified according to the new write
API and new tests have been added. This commit improves the current
tests and adds another test as well which is failing at the
moment. The commit saves the progress.