raft: Better wording for timeout range and seed
The current wording was a little confusing and it took me, the author
of this code myself longer than I had expected to understand what was
going on with all the TIMEOUT_SEED and TIMEOUT_RANGE
hullabaloo. Tweaking the wording a little appears to make that easier.

Well, until I have to revisit this piece of code again that is and
spend time understanding it. :D

Only time will tell. ;)