test: Remove test_write_to_candidate_with_leader
This test was fundamentally incorrect because here's what I was
trying:

- Elect a leader in term 1.
- Timeout a follower so that it becomes the candidate in term 2.
- Send write requests to this node.

At this point I was expecting it to respond as a candidate, however
the node becomes a leader in the meantime.

Note: There are calls to `timer:sleep(1)` explicitly in the tests so
that once a node has timedout it can have enough time to complete a
leader election before proceeding forward with the test. However, I've
noticed that in the past not using the sleep makes the tests fail as
the leader election is yet to be completed, thus throwing off the entire
expected state before running the tests. Similarly, in this test after
having promoted a node to candidate in the second term, the candidate
should stay put instead of becoming the leader. This remains a point
of confusion.

One chain of thought is that something changed between OTP 20 and OTP
21. Experimenting with removing all the `timer:sleep(1)` from test
maybe a good next step, but let us keep that for another commit! :)